+++ platformio.ini
@@ -55,6 +55,16 @@ build_flags =
   -DRX_ELRS_GENERIC_900_ESP8285
 
 
+[env:rx-generic-900-pa]
+extends = env_common_esp82xx
+board = esp8285
+build_src_filter = 
+  ${env_common_esp82xx.build_src_filter} 
+  +<modules/sx12xx-lib/src/sx127x.cpp>
+build_flags = 
+  -DRX_ELRS_GENERIC_900_PA_ESP8285
+
+
 [env:rx-generic-2400]
 extends = env_common_esp82xx
 board = esp8285
@@ -77,14 +87,10 @@ build_flags =
 
 ; ELRS boards
 
-[env:rx-bayckrc-nano-pro-900]
-extends = env_common_esp82xx
-board = esp8285
-build_src_filter = 
-  ${env_common_esp82xx.build_src_filter} 
-  +<modules/sx12xx-lib/src/sx127x.cpp>
+[env:rx-bayck-nano-pro-900]
+extends = env:rx-generic-900-pa
 build_flags = 
-  -DRX_ELRS_BAYCKRC_NANO_PRO_900_ESP8285
+  -DRX_ELRS_BAYCK_NANO_PRO_900_ESP8285
 
 
 [env:rx-speedybee-nano-2400]
